fastApiProject/requirements.txt
fastapi~=0.115.4
pydantic~=2.9.2
pandas~=2.2.3
PyYAML~=6.0.2
numpy~=1.26.4
keras~=3.6.0
scikit-learn~=1.5.2
joblib~=1.4.2
nltk~=3.9.1
packaging~=24.1